fix: resolve AI review comments, fix all-features build, dedupe benches
- builder.rs: fix broken intra-doc link (with_connection_strategy) that
  failed check-doc; re-add `redis::IntoConnectionInfo` import gated on the
  `embedded` feature (a prior commit removed it to silence an unused-import
  warning, which broke the embedded path under --all-features).
- async_strategies / resource_usage benches: propagate query and join
  errors (expect) instead of swallowing them, so failed iterations can't
  skew measurements.
- multiplexed_async example: surface JoinError instead of dropping it.
- embedded_integration: single-socket multiplexed test now fails fast
  after prerequisites are confirmed instead of silently skipping.
- benches/common: extract shared connection_info/build_client/counts to
  remove duplication between the two benchmarks.

Validated: cargo fmt --check, cargo doc --all, clippy (default + all
features), cargo-deny, multiplexing_parity (5/5), embedded compile, and
the full lib suite (307 pass) against a populated server.
