+++ examples/multiplexed_async.rs
@@ -55,8 +55,7 @@ async fn main() -> FalkorResult<()> {
 
     let mut completed = 0;
     while let Some(joined) = join_set.join_next().await {
-        let result = joined.expect("task should not panic");
-        let (_i, _value) = result?;
+        let (_i, _value) = joined.expect("task should not panic")?;
         completed += 1;
     }
     println!("completed {completed} concurrent queries over 4 multiplexed sockets");
